modifying some stuff in menu

+++ Assets/Scripts/Delivery Game Scripts/Manager.cs
@@ -3,8 +3,8 @@ public static class Manager
 {
     public static bool  gameStarted = false;
     public static bool  choosingCar = false;
-    public static float dynamicLatitude = .0f;
-    public static float dynamicLongitude = .0f;
+    public static double dynamicLatitude = .0f;
+    public static double dynamicLongitude = .0f;
     public static bool  locationQueryComplete = false;
     public static int   selectedVehicle = 0;
